WvStreams: WSID (WvStream ID) type was only 16 bits, which could easily get exceeded in a leaky, long-running program; extended to 32 bits.
git-svn-id: file:///home/apenwarr/alumnit-svn/public/branches/fig@10981 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ include/iwvstream.h
@@ -24,7 +24,7 @@ class WvStream;
 
 /* The stream gets passed back as a parameter. */
 typedef WvCallback<void, WvStream&> IWvStreamCallback;
-typedef unsigned short WSID;
+typedef unsigned int WSID;
 
 class IWvStream : public WvErrorBase, public IObject
 {
